Refactors project structure
