+++ src/gateway/gateway.cpp
