+++ src/gateway/gateway.h
