+++ src/gateway/socket.cpp
