+++ src/util/linked_list.h
