+++ tests/helpers.h
@@ -1,7 +1,7 @@
 #ifndef HELPERS_H
 #define HELPERS_H
 
-#include "../order_book/order_book.h"
+#include "../src/order_book/order_book.h"
 
 Order *createDefaultOrder();
 Order *customOrder(int price, int quantity, char side);
