+++ tests/order_book/book.test.cpp
@@ -1,8 +1,8 @@
 #include <catch2/catch_all.hpp>
 #include <catch2/catch_test_macros.hpp>
 
-#include "../../order_book/book.h"
-#include "../../order_book/order_book.h"
+#include "../../src/order_book/book.h"
+#include "../../src/order_book/order_book.h"
 #include "../helpers.h"
 
 TEST_CASE("Book test - Add & cancel order") {
