+++ tests/util/disruptor.test.cpp
@@ -1,7 +1,7 @@
 #include <catch2/catch_all.hpp>
 #include <catch2/catch_test_macros.hpp>
 
-#include "../../util/disruptor.h"
+#include "../../src/util/disruptor.h"
 
 const char *disruptor_pool_name = "/disruptor_pool";
